Removed okhttp dependency
- Removed okhttp dependency
- Upgraded to Java 21
- Simplified OUIDBDownloader code
- Split Downloader and Parser
