+++ Dockerfile
@@ -1,5 +1,5 @@
 # base build image
-FROM maven:3.6-openjdk-17 AS maven
+FROM maven:3.9.9-eclipse-temurin-21-jammy AS maven
 
 # copy the project files
 COPY ./pom.xml ./pom.xml
@@ -14,7 +14,7 @@ COPY ./src ./src
 RUN mvn package --batch-mode --quiet -Dmaven.test.skip=true -Dmaven.javadoc.skip=true -Djacoco.skip=true
 
 # final base image
-FROM eclipse-temurin:17-jre-jammy
+FROM eclipse-temurin:21-jre-jammy
 
 # Build-time metadata as defined at https://github.com/opencontainers/image-spec/blob/master/annotations.md
 LABEL org.opencontainers.image.title="OUIDB to JSON publisher" \
@@ -24,4 +24,4 @@ LABEL org.opencontainers.image.title="OUIDB to JSON publisher" \
 COPY --from=maven target/ouidb-to-json-publisher-jar-with-dependencies.jar /
 ENV DATA='/var/data'
 
-CMD java -jar /ouidb-to-json-publisher-jar-with-dependencies.jar
+CMD ["java", "-jar",  "/ouidb-to-json-publisher-jar-with-dependencies.jar"]
